refactor(demo): split code in features
